小説保存フォルダ初期化時に必須ファイルがすでにある場合は上書きしない

+++ lib/template.rb
@@ -15,12 +15,16 @@ class Template
   # src_filename  読み込みたいテンプレートファイル名(.erb は省略する)
   # dest_filepath 保存先ファイルパス。ディレクトリならファイル名はsrcと同じ名前で保存する
   # _binding      変数とか設定したいスコープの binding 変数を渡す
+  # overwrite     上書きするか
   #
-  def self.write(src_filename, dest_filepath, _binding)
-    result = get(src_filename, _binding)
+  def self.write(src_filename, dest_filepath, _binding, overwrite = false)
     if File.directory?(dest_filepath)
       dest_filepath = File.join(dest_filepath, src_filename)
     end
+    unless overwrite
+      return if File.exists?(dest_filepath)
+    end
+    result = get(src_filename, _binding)
     File.write(dest_filepath, result)
   end
